Successfully refactored spotify route! Please change your .env file to include /spotify/callback in the route.
